requirements.txt
asgiref==3.5.1
astroid==2.4.2
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
Django==4.0.4
djangorestframework==3.13.1
graphviz==0.8.4
idna==2.6
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.1
jupyter-client==6.1.3
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
mxnet==1.5.0
networkx==2.4
numpy==1.18.2
parso==0.7.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
protobuf==3.6.1
Pygments==2.6.1
pylint==2.5.3
pytz==2022.1
pywin32==228
PyYAML==5.3.1
pyzmq==19.0.1
six==1.14.0
sqlparse==0.4.2
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
tzdata==2022.1
urllib3==1.22
wcwidth==0.2.5
wincertstore==0.2